feat: ensure /cql and /object endpoints work with GeoJSON; include counts with /cql GeoJSON.
