+++ prez/renderers/renderer.py
@@ -2,20 +2,47 @@ import io
 import json
 import logging
 import time
-
+from datetime import datetime
+from typing import Dict, Optional
+from urllib.parse import urlencode
+from zoneinfo import ZoneInfo
+from httpx import URL
+from aiocache import cached
+from fastapi import Depends
 from fastapi import status
 from fastapi.exceptions import HTTPException
 from fastapi.responses import StreamingResponse
 from rdf2geojson import convert
-from rdflib import RDF, Graph, URIRef
+from rdflib import Graph, Literal
+from rdflib import URIRef
+from rdflib.namespace import GEO, RDF
+from sparql_grammar_pydantic import (
+    IRI,
+    GraphPatternNotTriples,
+    GroupGraphPattern,
+    GroupGraphPatternSub,
+    OptionalGraphPattern,
+    TriplesBlock,
+    TriplesSameSubject,
+    TriplesSameSubjectPath,
+    Var,
+)
 
+from prez.cache import endpoints_graph_cache
 from prez.cache import prefix_graph
+from prez.config import settings
+from prez.dependencies import get_endpoint_uri, get_system_repo, get_url
+from prez.models.ogc_features import Collection, Collections, Link, Links, Queryables
+from prez.models.query_params import QueryParams
+from prez.reference_data.prez_ns import OGCFEAT, ONT, PREZ
 from prez.renderers.csv_renderer import render_csv_dropdown
 from prez.renderers.json_renderer import NotFoundError, render_json_dropdown
 from prez.repositories import Repo
 from prez.services.annotations import get_annotation_properties
-from prez.services.connegp_service import RDF_MEDIATYPES, RDF_SERIALIZER_TYPES_MAP
+from prez.services.connegp_service import RDF_MEDIATYPES
+from prez.services.connegp_service import RDF_SERIALIZER_TYPES_MAP
 from prez.services.curie_functions import get_curie_id_for_uri
+from prez.services.query_generation.shacl import NodeShape
 
 log = logging.getLogger(__name__)
 
@@ -28,6 +55,8 @@ async def return_from_graph(
     selected_class: URIRef,
     repo: Repo,
     system_repo: Repo,
+    query_params: Optional[QueryParams] = None,
+    url: str = None
 ):
     profile_headers["Content-Disposition"] = "inline"
 
@@ -66,6 +95,12 @@ async def return_from_graph(
 
     elif str(mediatype) == "application/geo+json":
         geojson = convert(g=graph, do_validate=False, iri2id=get_curie_id_for_uri)
+        count = None  # for an object; no count query.
+        s_o = graph.subject_objects(predicate=PREZ["count"])  # for a list; graph will contain a count.
+        for tup in s_o:
+            str_count = str(tup[1])
+            count = get_geojson_int_count(str_count)
+        headers, geojson = await generate_geojson_extras(count, geojson, query_params, "application/geo+json", url)
         content = io.BytesIO(json.dumps(geojson).encode("utf-8"))
         return StreamingResponse(content=content, media_type=mediatype)
 
@@ -85,6 +120,15 @@ async def return_from_graph(
         raise HTTPException(
             status.HTTP_400_BAD_REQUEST, f"Unsupported mediatype: {mediatype}."
         )
+
+def get_geojson_int_count(count_str: str):
+    if count_str.startswith(">"):
+        count = int(
+            count_str[1:]
+        )  # TODO increment maximum counts based on current page.
+    else:
+        count = int(count_str)
+    return count
 
 
 async def return_rdf(graph: Graph, mediatype, profile_headers):
@@ -112,3 +156,226 @@ async def return_annotated_rdf(
     log.debug(f"Time to get annotations: {time.time() - t_start}")
     # return graph.__iadd__(annotations_graph)
     return annotations_graph
+
+
+async def generate_geojson_extras(count, geojson, query_params, selected_mediatype, url):
+    all_links = create_self_alt_links(selected_mediatype, url, query_params, count)
+    all_links_dict = Links(links=all_links).model_dump(exclude_none=True)
+    link_headers = generate_link_headers(all_links)
+    geojson["links"] = all_links_dict["links"]
+    geojson["timeStamp"] = get_brisbane_timestamp()
+    if count:  # listing
+        geojson["numberMatched"] = count
+        geojson["numberReturned"] = len(geojson["features"])
+    return link_headers, geojson
+
+
+def _add_inbound_triple_pattern_match(construct_tss_list):
+    triple = (Var(value="inbound_s"), Var(value="inbound_p"), Var(value="focus_node"))
+    construct_tss_list.append(TriplesSameSubject.from_spo(*triple))
+    inbound_tssp_list = [TriplesSameSubjectPath.from_spo(*triple)]
+    opt_inbound_gpnt = GraphPatternNotTriples(
+        content=OptionalGraphPattern(
+            group_graph_pattern=GroupGraphPattern(
+                content=GroupGraphPatternSub(
+                    triples_block=TriplesBlock.from_tssp_list(inbound_tssp_list)
+                )
+            )
+        )
+    )
+    return opt_inbound_gpnt
+
+
+def create_collections_json(
+    item_graph, annotations_graph, url, selected_mediatype, query_params, count: str
+):
+    collections_list = []
+    for s, p, o in item_graph.triples((None, RDF.type, GEO.FeatureCollection)):
+        curie_id = get_curie_id_for_uri(s)
+        collections_list.append(
+            Collection(
+                id=curie_id,
+                title=annotations_graph.value(
+                    subject=s, predicate=PREZ.label, default=None
+                ),
+                description=annotations_graph.value(
+                    subject=s, predicate=PREZ.description, default=None
+                ),
+                links=[
+                    Link(
+                        href=URIRef(
+                            f"{settings.system_uri}{url.path}/{curie_id}/items?{urlencode({'_mediatype': mt})}"
+                        ),
+                        rel="items",
+                        type=mt,
+                    )
+                    for mt in ["application/geo+json", *RDF_MEDIATYPES]
+                ],
+            )
+        )
+    self_alt_links = create_self_alt_links(selected_mediatype, url, query_params, count)
+    collections = Collections(
+        collections=collections_list,
+        links=self_alt_links,
+    )
+    return collections
+
+
+def create_self_alt_links(selected_mediatype, url, query_params=None, count=None):
+    self_alt_links = []
+    for mt in [selected_mediatype, *RDF_MEDIATYPES]:
+        self_alt_links.append(
+            Link(
+                href=URIRef(
+                    f"{settings.system_uri}{url.path}?{urlencode( dict(URL(str(url)).params) | {'_mediatype': mt} )}"
+                ),
+                rel="self" if mt == selected_mediatype else "alternate",
+                type=mt,
+                title="this document",
+            )
+        )
+    if count:  # only for listings; add prev/next links
+        page = query_params.page
+        limit = query_params.limit
+        if page != 1:
+            prev_page = page - 1
+            self_alt_links.append(
+                Link(
+                    href=URIRef(
+                        f"{settings.system_uri}{url.path}?{urlencode({'_mediatype': selected_mediatype, 'page': prev_page, 'limit': limit})}"
+                    ),
+                    rel="prev",
+                    type=selected_mediatype,
+                    title="previous page",
+                )
+            )
+        if count > page * limit:
+            next_page = page + 1
+            self_alt_links.append(
+                Link(
+                    href=URIRef(
+                        f"{settings.system_uri}{url.path}?{urlencode({'_mediatype': selected_mediatype, 'page': next_page, 'limit': limit})}"
+                    ),
+                    rel="next",
+                    type=selected_mediatype,
+                    title="next page",
+                )
+            )
+    return self_alt_links
+
+
+def generate_link_headers(links) -> Dict[str, str]:
+    link_header = ", ".join(
+        [f'<{link.href}>; rel="{link.rel}"; type="{link.type}"' for link in links]
+    )
+    return {"Link": link_header}
+
+
+async def handle_alt_profile(original_endpoint_type, pmts):
+    endpoint_nodeshape_map = {
+        ONT["ObjectEndpoint"]: URIRef("http://example.org/ns#AltProfilesForObject"),
+        ONT["ListingEndpoint"]: URIRef("http://example.org/ns#AltProfilesForListing"),
+    }
+    endpoint_uri = endpoint_nodeshape_map[original_endpoint_type]
+    endpoint_nodeshape = NodeShape(
+        uri=endpoint_uri,
+        graph=endpoints_graph_cache,
+        kind="endpoint",
+        focus_node=Var(value="focus_node"),
+        path_nodes={
+            "path_node_1": IRI(value=pmts.selected["class"])
+        },  # hack - not sure how (or if) the class can be
+        # 'dynamicaly' expressed in SHACL. The class is only known at runtime
+    )
+    return endpoint_nodeshape
+
+
+def get_brisbane_timestamp():
+    # Get current time in Brisbane
+    brisbane_time = datetime.now(ZoneInfo("Australia/Brisbane"))
+
+    # Format the timestamp
+    timestamp = brisbane_time.strftime("%Y-%m-%dT%H:%M:%S%z")
+
+    # Insert colon in timezone offset
+    return f"{timestamp[:-2]}:{timestamp[-2:]}"
+
+
+# TODO cache this
+async def generate_queryables_from_shacl_definition(
+    url: str = Depends(get_url),
+    endpoint_uri: URIRef = Depends(get_endpoint_uri),
+    system_repo: Repo = Depends(get_system_repo),
+):
+    query = """
+    PREFIX cql: <http://www.opengis.net/doc/IS/cql2/1.0/>
+    PREFIX dcterms: <http://purl.org/dc/terms/>
+    PREFIX sh: <http://www.w3.org/ns/shacl#>
+    PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
+    CONSTRUCT {
+    ?queryable cql:id ?id ;
+    	cql:name ?title ;
+    	cql:description ?description ;
+    	cql:datatype ?type ;
+    	cql:enum ?enums .
+    }
+    WHERE {?queryable a cql:Queryable ;
+        dcterms:identifier ?id ;
+        sh:name ?title ;
+        sh:description ?description ;
+        sh:datatype ?type .
+        OPTIONAL { ?queryable sh:in/rdf:rest*/rdf:first ?enums }
+    }
+    """
+    g, _ = await system_repo.send_queries([query], [])
+    if (
+        len(g) == 0
+    ):  # will auto generate queryables from data - less preferable approach
+        return None
+    jsonld_string = g.serialize(format="json-ld")
+    jsonld = json.loads(jsonld_string)
+    queryable_props = {}
+    for item in jsonld:
+        id_value = item["http://www.opengis.net/doc/IS/cql2/1.0/id"][0]["@value"]
+        queryable_props[id_value] = {
+            "title": item["http://www.opengis.net/doc/IS/cql2/1.0/name"][0]["@value"],
+            "type": item["http://www.opengis.net/doc/IS/cql2/1.0/datatype"][0][
+                "@id"
+            ].split("#")[
+                -1
+            ],  # hack
+            "description": item["http://www.opengis.net/doc/IS/cql2/1.0/description"][
+                0
+            ]["@value"],
+        }
+        enum = item.get(
+            "http://www.opengis.net/doc/IS/cql2/1.0/enum"
+        )  # enums are optional.
+        if enum:
+            queryable_props[id_value]["enum"] = [enum_item["@id"] for enum_item in enum]
+    if endpoint_uri == OGCFEAT["queryables-global"]:
+        title = "Global Queryables"
+        description = (
+            "Global queryable properties for all collections in the OGC Features API."
+        )
+    else:
+        title = "Local Queryables"
+        description = (
+            "Local queryable properties for the collection in the OGC Features API."
+        )
+    queryable_params = {
+        "$id": f"{settings.system_uri}{url.path}",
+        "title": title,
+        "description": description,
+        "properties": queryable_props,
+    }
+    return Queryables(**queryable_params)
+
+
+@cached(ttl=600, key=lambda collection_uri: collection_uri)
+async def _cached_feature_collection_query(
+    collection_uri, data_repo, feature_collection_query
+):
+    """cache the feature collection information for 10 minutes as it is an expensive query at present"""
+    fc_item_graph, _ = await data_repo.send_queries([feature_collection_query], [])
+    return fc_item_graph
