+++ prez/routers/custom_endpoints.py
@@ -15,7 +15,7 @@ from prez.dependencies import (
     get_endpoint_structure,
     get_negotiated_pmts,
     get_profile_nodeshape,
-    get_system_repo,
+    get_system_repo, get_url,
 )
 from prez.models.query_params import QueryParams
 from prez.reference_data.prez_ns import ONT
@@ -51,6 +51,8 @@ def create_dynamic_route_handler(route_type: str):
             ),
             data_repo: Repo = Depends(get_data_repo),
             system_repo: Repo = Depends(get_system_repo),
+            url: str = Depends(get_url),
+
         ):
             return await listing_function(
                 data_repo=data_repo,
@@ -64,6 +66,7 @@ def create_dynamic_route_handler(route_type: str):
                 profile_nodeshape=profile_nodeshape,
                 query_params=query_params,
                 original_endpoint_type=ONT["ListingEndpoint"],
+                url=url
             )
 
         return dynamic_list_handler
@@ -75,6 +78,7 @@ def create_dynamic_route_handler(route_type: str):
             profile_nodeshape: NodeShape = Depends(get_profile_nodeshape),
             data_repo: Repo = Depends(get_data_repo),
             system_repo: Repo = Depends(get_system_repo),
+            url: str = Depends(get_url),
         ):
             return await object_function(
                 data_repo=data_repo,
@@ -82,6 +86,7 @@ def create_dynamic_route_handler(route_type: str):
                 endpoint_structure=endpoint_structure,
                 pmts=pmts,
                 profile_nodeshape=profile_nodeshape,
+                url=url
             )
 
         return dynamic_object_handler
