+++ prez/services/listings.py
@@ -2,13 +2,7 @@ import copy
 import io
 import json
 import logging
-from datetime import datetime
-from typing import Dict
-from urllib.parse import urlencode
-from zoneinfo import ZoneInfo
 
-from aiocache import cached
-from fastapi import Depends
 from fastapi.responses import PlainTextResponse
 from rdf2geojson import convert
 from rdflib import Literal, URIRef
@@ -18,26 +12,19 @@ from sparql_grammar_pydantic import (
     Bind,
     Expression,
     GraphPatternNotTriples,
-    GroupGraphPattern,
-    GroupGraphPatternSub,
     IRIOrFunction,
-    OptionalGraphPattern,
     PrimaryExpression,
-    TriplesBlock,
     TriplesSameSubject,
     TriplesSameSubjectPath,
     Var,
 )
 
 from prez.cache import endpoints_graph_cache
-from prez.config import settings
-from prez.dependencies import get_endpoint_uri, get_system_repo, get_url
 from prez.enums import NonAnnotatedRDFMediaType
-from prez.models.ogc_features import Collection, Collections, Link, Links, Queryables
-from prez.reference_data.prez_ns import ALTREXT, OGCFEAT, ONT, PREZ
-from prez.renderers.renderer import return_annotated_rdf, return_from_graph
-from prez.repositories import Repo
-from prez.services.connegp_service import RDF_MEDIATYPES
+from prez.reference_data.prez_ns import ALTREXT, OGCFEAT, PREZ
+from prez.renderers.renderer import return_annotated_rdf, return_from_graph, generate_geojson_extras, \
+    handle_alt_profile, generate_queryables_from_shacl_definition, _add_inbound_triple_pattern_match, \
+    _cached_feature_collection_query, create_collections_json, generate_link_headers, get_geojson_int_count
 from prez.services.curie_functions import get_curie_id_for_uri
 from prez.services.generate_queryables import generate_queryables_json
 from prez.services.link_generation import add_prez_links
@@ -53,6 +40,14 @@ log = logging.getLogger(__name__)
 DWC = Namespace("http://rs.tdwg.org/dwc/terms/")
 
 
+def _add_geom_triple_pattern_match(tssp_list: list[TriplesSameSubjectPath]):
+    triples = [
+        (Var(value="focus_node"), IRI(value=GEO.hasGeometry),Var(value="bn")),
+        (Var(value="bn"), IRI(value=GEO.asWKT), Var(value="wkt")),
+    ]
+    tssp_list.extend([TriplesSameSubjectPath.from_spo(*triple) for triple in triples])
+
+
 async def listing_function(
     data_repo,
     system_repo,
@@ -65,6 +60,7 @@ async def listing_function(
     profile_nodeshape,
     query_params,
     original_endpoint_type,
+    url
 ):
     if (
         pmts.selected["profile"] == ALTREXT["alt-profile"]
@@ -78,6 +74,9 @@ async def listing_function(
         concept_hierarchy_query=concept_hierarchy_query,
         query_params=query_params,
     )
+    if pmts.selected["mediatype"] == "application/geo+json":  # Ensure the focus nodes have a geometry in the SPARQL
+        # subselect. If they are missing, the subsequent GeoJSON conversion will drop any Features without geometries.
+        _add_geom_triple_pattern_match(subselect_kwargs["inner_select_tssp_list"])
 
     # merge subselect and profile triples same subject (for construct triples)
     construct_tss_list = []
@@ -113,7 +112,8 @@ async def listing_function(
         return PlainTextResponse(queries[0], media_type="application/sparql-query")
 
     # add a count query if it's an annotated mediatype
-    if "anot+" in pmts.selected["mediatype"] and not search_query:
+    if ("anot+" in pmts.selected["mediatype"] and not search_query) or \
+            (pmts.selected["mediatype"] == "application/geo+json"):
         subselect = copy.deepcopy(main_query.inner_select)
         count_query = CountQuery(original_subselect=subselect).to_string()
         queries.append(count_query)
@@ -144,6 +144,8 @@ async def listing_function(
         pmts.selected["class"],
         data_repo,
         system_repo,
+        query_params,
+        url
     )
 
 
@@ -215,12 +217,23 @@ async def ogc_features_listing_function(
                 **subselect_kwargs,
             ).to_string()
             queries.append(query)
-    elif not collection_uri:  # list Feature Collections
-        query = PrezQueryConstructor(
-            construct_tss_list=construct_tss_list,
-            profile_triples=profile_nodeshape.tssp_list,
-            **subselect_kwargs,
-        )
+    elif not collection_uri:  # list Feature Collections OR Features within a Feature Collection
+        if any([query_params._filter, query_params.datetime, query_params.bbox]):  # list Features within a Feature Collection
+            # create the tssp for a Feature rather than FC
+            feat_path_nodes = {k: v for k, v in endpoint_nodeshape.path_nodes.items() if k == "path_node_2"}
+            feat_ep_ns = NodeShape(
+                uri=URIRef("http://example.org/ns#Feature"),
+                graph=endpoints_graph_cache,
+                kind="endpoint",
+                focus_node=Var(value="focus_node"),
+                path_nodes=feat_path_nodes
+            )
+        else:
+            query = PrezQueryConstructor(
+                construct_tss_list=construct_tss_list,
+                profile_triples=profile_nodeshape.tssp_list,
+                **subselect_kwargs,
+            )
         queries.append(query.to_string())
         # add the count query
         subselect = copy.deepcopy(query.inner_select)
@@ -285,12 +298,8 @@ async def ogc_features_listing_function(
         count_g, _ = await data_repo.send_queries([count_query], [])
         if count_g:
             count = str(next(iter(count_g.objects())))
-            if count.startswith(">"):
-                count = int(
-                    count[1:]
-                )  # TODO increment maximum counts based on current page.
-            else:
-                count = int(count)
+            count = get_geojson_int_count(count)
+
 
     if selected_mediatype == "application/json":
         if endpoint_uri_type[0] in [
@@ -327,227 +336,12 @@ async def ogc_features_listing_function(
 
     elif selected_mediatype == "application/geo+json":
         geojson = convert(g=item_graph, do_validate=False, iri2id=get_curie_id_for_uri)
-        all_links = create_self_alt_links(selected_mediatype, url, query_params, count)
-        all_links_dict = Links(links=all_links).model_dump(exclude_none=True)
-        link_headers = generate_link_headers(all_links)
-        geojson["links"] = all_links_dict["links"]
-        geojson["timeStamp"] = get_brisbane_timestamp()
-        geojson["numberMatched"] = count
-        geojson["numberReturned"] = len(geojson["features"])
+        link_headers, geojson = await generate_geojson_extras(
+            count, geojson, query_params, selected_mediatype, url
+        )
         content = io.BytesIO(json.dumps(geojson).encode("utf-8"))
     elif selected_mediatype in NonAnnotatedRDFMediaType:
         content = io.BytesIO(
             item_graph.serialize(format=selected_mediatype, encoding="utf-8")
         )
     return content, link_headers
-
-
-def _add_inbound_triple_pattern_match(construct_tss_list):
-    triple = (Var(value="inbound_s"), Var(value="inbound_p"), Var(value="focus_node"))
-    construct_tss_list.append(TriplesSameSubject.from_spo(*triple))
-    inbound_tssp_list = [TriplesSameSubjectPath.from_spo(*triple)]
-    opt_inbound_gpnt = GraphPatternNotTriples(
-        content=OptionalGraphPattern(
-            group_graph_pattern=GroupGraphPattern(
-                content=GroupGraphPatternSub(
-                    triples_block=TriplesBlock.from_tssp_list(inbound_tssp_list)
-                )
-            )
-        )
-    )
-    return opt_inbound_gpnt
-
-
-def create_collections_json(
-    item_graph, annotations_graph, url, selected_mediatype, query_params, count: str
-):
-    collections_list = []
-    for s, p, o in item_graph.triples((None, RDF.type, GEO.FeatureCollection)):
-        curie_id = get_curie_id_for_uri(s)
-        collections_list.append(
-            Collection(
-                id=curie_id,
-                title=annotations_graph.value(
-                    subject=s, predicate=PREZ.label, default=None
-                ),
-                description=annotations_graph.value(
-                    subject=s, predicate=PREZ.description, default=None
-                ),
-                links=[
-                    Link(
-                        href=URIRef(
-                            f"{settings.system_uri}{url.path}/{curie_id}/items?{urlencode({'_mediatype': mt})}"
-                        ),
-                        rel="items",
-                        type=mt,
-                    )
-                    for mt in ["application/geo+json", *RDF_MEDIATYPES]
-                ],
-            )
-        )
-    self_alt_links = create_self_alt_links(selected_mediatype, url, query_params, count)
-    collections = Collections(
-        collections=collections_list,
-        links=self_alt_links,
-    )
-    return collections
-
-
-def create_self_alt_links(selected_mediatype, url, query_params=None, count=None):
-    self_alt_links = []
-    for mt in [selected_mediatype, *RDF_MEDIATYPES]:
-        self_alt_links.append(
-            Link(
-                href=URIRef(
-                    f"{settings.system_uri}{url.path}?{urlencode({'_mediatype': mt})}"
-                ),
-                rel="self" if mt == selected_mediatype else "alternate",
-                type=mt,
-                title="this document",
-            )
-        )
-    if count:  # only for listings; add prev/next links
-        page = query_params.page
-        limit = query_params.limit
-        if page != 1:
-            prev_page = page - 1
-            self_alt_links.append(
-                Link(
-                    href=URIRef(
-                        f"{settings.system_uri}{url.path}?{urlencode({'_mediatype': selected_mediatype, 'page': prev_page, 'limit': limit})}"
-                    ),
-                    rel="prev",
-                    type=selected_mediatype,
-                    title="previous page",
-                )
-            )
-        if count > page * limit:
-            next_page = page + 1
-            self_alt_links.append(
-                Link(
-                    href=URIRef(
-                        f"{settings.system_uri}{url.path}?{urlencode({'_mediatype': selected_mediatype, 'page': next_page, 'limit': limit})}"
-                    ),
-                    rel="next",
-                    type=selected_mediatype,
-                    title="next page",
-                )
-            )
-    return self_alt_links
-
-
-def generate_link_headers(links) -> Dict[str, str]:
-    link_header = ", ".join(
-        [f'<{link.href}>; rel="{link.rel}"; type="{link.type}"' for link in links]
-    )
-    return {"Link": link_header}
-
-
-async def handle_alt_profile(original_endpoint_type, pmts):
-    endpoint_nodeshape_map = {
-        ONT["ObjectEndpoint"]: URIRef("http://example.org/ns#AltProfilesForObject"),
-        ONT["ListingEndpoint"]: URIRef("http://example.org/ns#AltProfilesForListing"),
-    }
-    endpoint_uri = endpoint_nodeshape_map[original_endpoint_type]
-    endpoint_nodeshape = NodeShape(
-        uri=endpoint_uri,
-        graph=endpoints_graph_cache,
-        kind="endpoint",
-        focus_node=Var(value="focus_node"),
-        path_nodes={
-            "path_node_1": IRI(value=pmts.selected["class"])
-        },  # hack - not sure how (or if) the class can be
-        # 'dynamicaly' expressed in SHACL. The class is only known at runtime
-    )
-    return endpoint_nodeshape
-
-
-def get_brisbane_timestamp():
-    # Get current time in Brisbane
-    brisbane_time = datetime.now(ZoneInfo("Australia/Brisbane"))
-
-    # Format the timestamp
-    timestamp = brisbane_time.strftime("%Y-%m-%dT%H:%M:%S%z")
-
-    # Insert colon in timezone offset
-    return f"{timestamp[:-2]}:{timestamp[-2:]}"
-
-
-# TODO cache this
-async def generate_queryables_from_shacl_definition(
-    url: str = Depends(get_url),
-    endpoint_uri: URIRef = Depends(get_endpoint_uri),
-    system_repo: Repo = Depends(get_system_repo),
-):
-    query = """
-    PREFIX cql: <http://www.opengis.net/doc/IS/cql2/1.0/>
-    PREFIX dcterms: <http://purl.org/dc/terms/>
-    PREFIX sh: <http://www.w3.org/ns/shacl#>
-    PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-    CONSTRUCT {
-    ?queryable cql:id ?id ;
-    	cql:name ?title ;
-    	cql:description ?description ;
-    	cql:datatype ?type ;
-    	cql:enum ?enums .
-    }
-    WHERE {?queryable a cql:Queryable ;
-        dcterms:identifier ?id ;
-        sh:name ?title ;
-        sh:description ?description ;
-        sh:datatype ?type .
-        OPTIONAL { ?queryable sh:in/rdf:rest*/rdf:first ?enums }
-    }
-    """
-    g, _ = await system_repo.send_queries([query], [])
-    if (
-        len(g) == 0
-    ):  # will auto generate queryables from data - less preferable approach
-        return None
-    jsonld_string = g.serialize(format="json-ld")
-    jsonld = json.loads(jsonld_string)
-    queryable_props = {}
-    for item in jsonld:
-        id_value = item["http://www.opengis.net/doc/IS/cql2/1.0/id"][0]["@value"]
-        queryable_props[id_value] = {
-            "title": item["http://www.opengis.net/doc/IS/cql2/1.0/name"][0]["@value"],
-            "type": item["http://www.opengis.net/doc/IS/cql2/1.0/datatype"][0][
-                "@id"
-            ].split("#")[
-                -1
-            ],  # hack
-            "description": item["http://www.opengis.net/doc/IS/cql2/1.0/description"][
-                0
-            ]["@value"],
-        }
-        enum = item.get(
-            "http://www.opengis.net/doc/IS/cql2/1.0/enum"
-        )  # enums are optional.
-        if enum:
-            queryable_props[id_value]["enum"] = [enum_item["@id"] for enum_item in enum]
-    if endpoint_uri == OGCFEAT["queryables-global"]:
-        title = "Global Queryables"
-        description = (
-            "Global queryable properties for all collections in the OGC Features API."
-        )
-    else:
-        title = "Local Queryables"
-        description = (
-            "Local queryable properties for the collection in the OGC Features API."
-        )
-    queryable_params = {
-        "$id": f"{settings.system_uri}{url.path}",
-        "title": title,
-        "description": description,
-        "properties": queryable_props,
-    }
-    return Queryables(**queryable_params)
-
-
-@cached(ttl=600, key=lambda collection_uri: collection_uri)
-async def _cached_feature_collection_query(
-    collection_uri, data_repo, feature_collection_query
-):
-    """cache the feature collection information for 10 minutes as it is an expensive query at present"""
-    fc_item_graph, _ = await data_repo.send_queries([feature_collection_query], [])
-    return fc_item_graph
