+++ prez/services/objects.py
@@ -21,12 +21,12 @@ from prez.renderers.renderer import return_annotated_rdf, return_from_graph
 from prez.services.connegp_service import RDF_MEDIATYPES
 from prez.services.curie_functions import get_curie_id_for_uri
 from prez.services.link_generation import add_prez_links
-from prez.services.listings import (
+from prez.renderers.renderer import (
     create_self_alt_links,
     generate_link_headers,
-    get_brisbane_timestamp,
-    listing_function,
+    get_brisbane_timestamp
 )
+from prez.services.listings import listing_function
 from prez.services.query_generation.umbrella import PrezQueryConstructor
 
 log = logging.getLogger(__name__)
@@ -38,18 +38,12 @@ async def object_function(
         endpoint_structure,
         pmts,
         profile_nodeshape,
+        url
 ):
     if pmts.selected["profile"] == ALTREXT["alt-profile"]:
-        none_keys = [
-            "endpoint_nodeshape",
-            "concept_hierarchy_query",
-            "cql_parser",
-            "search_query",
-        ]
-        none_kwargs = {key: None for key in none_keys}
         query_params = QueryParams(
             mediatype=pmts.selected["mediatype"],
-            filter=None,
+            _filter=None,
             q=None,
             page=1,
             limit=100,
@@ -68,7 +62,11 @@ async def object_function(
             profile_nodeshape=profile_nodeshape,
             query_params=query_params,
             original_endpoint_type=ONT["ObjectEndpoint"],
-            **none_kwargs,
+            url=url,
+            endpoint_nodeshape=None,
+            concept_hierarchy_query=None,
+            cql_parser=None,
+            search_query=None,
         )
     if "anot+" in pmts.selected["mediatype"]:
         profile_nodeshape.tss_list.append(
@@ -114,6 +112,7 @@ async def object_function(
         pmts.selected["class"],
         data_repo,
         system_repo,
+        url=url
     )
