+++ tests/test_ogc.py
@@ -66,12 +66,12 @@ def client(test_repo: Repo) -> TestClient:
         ),
         "errorconditions",
         "feature",
-        # pytest.param(
-        #     "features",
-        #     marks=pytest.mark.xfail(
-        #         reason="endpoint that causes an error in pytest works manually with the same data in Fuseki"
-        #     ),
-        # ),
+        pytest.param(
+            "features",
+            marks=pytest.mark.xfail(
+                reason="endpoint that causes an error in pytest works manually with the same data in Fuseki"
+            ),
+        ),
         "general",
         "landingpage",
     ],
